mrv-analytics-agent/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-cloud-aiplatform
pandas
scikit-learn
db-dtypes